[sourcekit] Sink UID handling into the API library
This fixes a layering violation where UID handling was provided by a
dependent library, creating a dependency cycle if you tried to link the
lower level libraries dynamically.
