+++ tools/SourceKit/tools/sourcekitd/bin/XPC/Service/XPCService.cpp
@@ -71,7 +71,7 @@ public:
 
 static SKUIDToUIDMap UIDMap;
 
-sourcekitd_uid_t sourcekitd::SKDUIDFromUIdent(UIdent UID) {
+static sourcekitd_uid_t xpcSKDUIDFromUIdent(UIdent UID) {
   if (void *Tag = UID.getTag())
     return reinterpret_cast<sourcekitd_uid_t>(Tag);
 
@@ -108,7 +108,7 @@ sourcekitd_uid_t sourcekitd::SKDUIDFromUIdent(UIdent UID) {
   return skduid;
 }
 
-UIdent sourcekitd::UIdentFromSKDUID(sourcekitd_uid_t SKDUID) {
+static UIdent xpcUIdentFromSKDUID(sourcekitd_uid_t SKDUID) {
   // This should be used only for debugging/logging purposes.
 
   UIdent UID = UIDMap.get(SKDUID);
@@ -341,6 +341,13 @@ static void fatal_error_handler(void *user_data, const std::string& reason,
 int main(int argc, const char *argv[]) {
   llvm::install_fatal_error_handler(fatal_error_handler, 0);
   sourcekitd::enableLogging("sourcekit-serv");
+  sourcekitd_set_uid_handlers(
+      ^sourcekitd_uid_t(const char *uidStr) {
+        return xpcSKDUIDFromUIdent(UIdent(uidStr));
+      },
+      ^const char *(sourcekitd_uid_t uid) {
+        return xpcUIdentFromSKDUID(uid).c_str();
+      });
   sourcekitd::initializeService(
       getRuntimeLibPath(), getDiagnosticDocumentationPath(), postNotification);
