+++ tools/SourceKit/tools/sourcekitd/lib/API/CMakeLists.txt
@@ -12,6 +12,7 @@ add_sourcekit_library(sourcekitdAPI
   sourcekitdAPI-Common.cpp
   TokenAnnotationsArray.cpp
   ExpressionTypeArray.cpp
+  UIDHandling.cpp
 )
 target_link_libraries(sourcekitdAPI PRIVATE
   swiftBasic
